M2 PR-3: move AST test builders into shared non-test astbuild.go
The hand-built AST node constructors (numExpr/strExpr/identExpr/numAnn/strAnn/
param/block/exprStmt/returnStmt/valDecl/funcExpr) lived in infer_func_test.go.
Move them into a non-test file (astbuild.go, package solver) so they are shared
across the package's test files and usable by non-test code, not confined to the
test binary.

- Builders now stamp a shared builderSpan() (the placeholder span a hand-built
  node carries) instead of the test-only testSpan(); testSpan() delegates to
  builderSpan() so inline-built and builder-built nodes carry the same span and
  the existing error-span assertions stay consistent.
- infer_func_test.go keeps only render() and the tests.

No behavior change; build, vet, and the full solver test suite stay green.
